+++ projects/igniteui-angular/src/lib/grids/grid/grid.component.spec.ts
@@ -8,6 +8,7 @@ import { By } from '@angular/platform-browser';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { IgxGridAPIService } from './grid-api.service';
 import { IgxGridComponent } from './grid.component';
+import { IgxRowComponent } from '../row.component';
 import { IgxGridTransaction, IGridEditEventArgs } from '../grid-base.component';
 import { IgxColumnComponent } from '../column.component';
 import { IForOfState } from '../../directives/for-of/for_of.directive';
@@ -50,7 +51,8 @@ describe('IgxGrid Component Tests', () => {
                     IgxGridTestComponent,
                     IgxGridMarkupDeclarationComponent,
                     IgxGridRemoteVirtualizationComponent,
-                    IgxGridRemoteOnDemandComponent
+                    IgxGridRemoteOnDemandComponent,
+                    IgxGridEmptyMessage100PercentComponent
                 ],
                 imports: [
                     NoopAnimationsModule, IgxGridModule]
@@ -273,12 +275,17 @@ describe('IgxGrid Component Tests', () => {
 
             const grid = fixture.componentInstance.grid;
             const gridBody = fixture.debugElement.query(By.css(TBODY_CLASS));
+            const domGrid = fixture.debugElement.query(By.css('igx-grid')).nativeElement;
+
+            // make sure default width/height are applied when there is no data
+            expect(domGrid.style.height).toBe('100%');
+            expect(domGrid.style.width).toBe('100%');
 
             // Check for loaded rows in grid's container
             fixture.componentInstance.generateData(30);
             fixture.detectChanges();
             tick(1000);
-            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBeGreaterThan(1000);
+            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBe(548);
 
             // Check for empty filter grid message and body less than 100px
             const columns = fixture.componentInstance.grid.columns;
@@ -286,13 +293,13 @@ describe('IgxGrid Component Tests', () => {
             fixture.detectChanges();
             tick(100);
             expect(gridBody.nativeElement.textContent).toEqual(grid.emptyFilteredGridMessage);
-            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBeLessThan(100);
+            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBe(548);
 
             // Clear filter and check if grid's body height is restored based on all loaded rows
             grid.clearFilter(columns[0].field);
             fixture.detectChanges();
             tick(100);
-            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBeGreaterThan(1000);
+            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBe(548);
 
             // Clearing grid's data and check for empty grid message
             fixture.componentInstance.clearData();
@@ -311,6 +318,11 @@ describe('IgxGrid Component Tests', () => {
             const grid = fixture.componentInstance.grid;
             const gridBody = fixture.debugElement.query(By.css(TBODY_CLASS));
             let loadingIndicator = gridBody.query(By.css('.igx-grid__loading'));
+            const domGrid = fixture.debugElement.query(By.css('igx-grid')).nativeElement;
+
+            // make sure default width/height are applied when there is no data
+            expect(domGrid.style.height).toBe('100%');
+            expect(domGrid.style.width).toBe('100%');
 
             expect(loadingIndicator).not.toBeNull();
             expect(gridBody.nativeElement.textContent).not.toEqual(grid.emptyFilteredGridMessage);
@@ -319,7 +331,7 @@ describe('IgxGrid Component Tests', () => {
             fixture.componentInstance.generateData(30);
             fixture.detectChanges();
             tick(1000);
-            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBeGreaterThan(1000);
+            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBe(548);
 
             loadingIndicator = gridBody.query(By.css('.igx-grid__loading'));
             expect(loadingIndicator).toBeNull();
@@ -330,13 +342,13 @@ describe('IgxGrid Component Tests', () => {
             fixture.detectChanges();
             tick(100);
             expect(gridBody.nativeElement.textContent).toEqual(grid.emptyFilteredGridMessage);
-            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBeLessThan(100);
+            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBe(548);
 
             // Clear filter and check if grid's body height is restored based on all loaded rows
             grid.clearFilter(columns[0].field);
             fixture.detectChanges();
             tick(100);
-            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBeGreaterThan(1000);
+            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBe(548);
 
             // Clearing grid's data and check for empty grid message
             fixture.componentInstance.clearData();
@@ -471,6 +483,87 @@ describe('IgxGrid Component Tests', () => {
             expect(colHeaders.length).toBeGreaterThan(0);
             expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBeGreaterThan(500);
         }));
+
+        it('should render empty message when grid height is 100%', fakeAsync(() => {
+            const fixture = TestBed.createComponent(IgxGridEmptyMessage100PercentComponent);
+            fixture.detectChanges();
+
+            const grid = fixture.componentInstance.grid;
+            const gridBody = fixture.debugElement.query(By.css(TBODY_CLASS));
+            const domGrid = fixture.debugElement.query(By.css('igx-grid')).nativeElement;
+
+            // make sure default width/height are applied when there is no data
+            expect(domGrid.style.height).toBe('100%');
+            expect(domGrid.style.width).toBe('100%');
+
+            expect(parseInt(window.getComputedStyle(gridBody.nativeElement).height, 10)).toBeGreaterThan(0);
+            expect(gridBody.nativeElement.innerText).toMatch(grid.emptyGridMessage);
+        }));
+    });
+
+    describe('IgxGrid - virtualization tests', () => {
+        configureTestSuite();
+        beforeEach(async(() => {
+            TestBed.configureTestingModule({
+                declarations: [
+                    IgxGridTestComponent
+                ],
+                imports: [
+                    NoopAnimationsModule, IgxGridModule]
+            }).compileComponents();
+        }));
+
+        it('should change chunk size for every record after enlarging the grid and the horizontal dirs are scrambled', async () => {
+            const fix = TestBed.createComponent(IgxGridTestComponent);
+            for (let i = 2; i < 100; i++) {
+                fix.componentInstance.data.push({ index: i, value: i, desc: i, detail: i });
+            }
+            fix.componentInstance.columns[0].width = '400px';
+            fix.componentInstance.columns[1].width = '400px';
+            fix.componentInstance.columns.push(
+                { field: 'desc', header: 'desc', dataType: 'number', width: '400px', hasSummary: false },
+                { field: 'detail', header: 'detail', dataType: 'number', width: '400px', hasSummary: false }
+            );
+            fix.detectChanges();
+            fix.componentInstance.grid.verticalScrollContainer.getVerticalScroll().scrollTop = 100;
+            await wait(100);
+            fix.detectChanges();
+            fix.componentInstance.grid.verticalScrollContainer.getVerticalScroll().scrollTop = 250;
+            await wait(100);
+            fix.detectChanges();
+            fix.componentInstance.grid.width = '1300px';
+            await wait(100);
+            fix.detectChanges();
+            const rows = fix.componentInstance.grid.rowList.toArray();
+            for (let i = 0; i < rows.length; i++) {
+                const row = rows[i] as IgxRowComponent<any>;
+                expect(row.cells.length).toEqual(4);
+            }
+        });
+
+        it('should not keep a cached-out template as master after column resizing', async() => {
+            const fix = TestBed.createComponent(IgxGridTestComponent);
+            for (let i = 2; i < 100; i++) {
+                fix.componentInstance.data.push({ index: i, value: i, desc: i, detail: i });
+            }
+            fix.componentInstance.columns[0].width = '400px';
+            fix.componentInstance.columns[1].width = '400px';
+            fix.componentInstance.columns.push(
+                { field: 'desc', header: 'desc', dataType: 'number', width: '400px', hasSummary: false },
+                { field: 'detail', header: 'detail', dataType: 'number', width: '400px', hasSummary: false }
+            );
+            fix.detectChanges();
+            fix.componentInstance.grid.groupBy({ fieldName: 'value', dir: SortingDirection.Asc });
+            fix.detectChanges();
+            fix.componentInstance.grid.getColumnByName('index').width = '100px';
+            fix.detectChanges();
+            await wait();
+            const rows = fix.componentInstance.grid.dataRowList.toArray();
+            for (let i = 0; i < rows.length; i++) {
+                const row = rows[i] as IgxRowComponent<any>;
+                expect(row.cells.length).toEqual(4);
+            }
+        });
     });
 
     describe('IgxGrid - default rendering for rows and columns', () => {
@@ -1076,6 +1169,77 @@ describe('IgxGrid Component Tests', () => {
                 const firstCellInputValue = firstRowCells[1].nativeElement.textContent.trim();
                 expect(firstCellInputValue).toEqual('4');
             }));
+
+        it(`GetNextCell: should return correctly next cell coordinates`, async() => {
+            const fix = TestBed.createComponent(IgxGridDefaultRenderingComponent);
+            fix.componentInstance.initColumnsRows(15, 5);
+            fix.detectChanges();
+
+            const grid = fix.componentInstance.grid;
+            grid.height = '500px';
+            await wait(30);
+            fix.detectChanges();
+
+            grid.getColumnByName('col2').editable = true;
+            fix.detectChanges();
+            grid.getColumnByName('col4').editable = true;
+            fix.detectChanges();
+            // when the next cell is on the same row
+            let nextCellCoords = grid.getNextCell(0, 0, (col) => col.editable);
+            expect(nextCellCoords).toEqual({rowIndex: 0, visibleColumnIndex: 2});
+            // when the next cell is on the next row
+            nextCellCoords = grid.getNextCell(0, 4, (col) => col.editable);
+            expect(nextCellCoords).toEqual({rowIndex: 1, visibleColumnIndex: 2});
+            // when the next cell is not in the view
+            nextCellCoords = grid.getNextCell(9, 4, (col) => col.editable);
+            expect(nextCellCoords).toEqual({rowIndex: 10, visibleColumnIndex: 2});
+            // when the current row and column index are not valid
+            nextCellCoords = grid.getNextCell(-10, 14, (col) => col.editable);
+            expect(nextCellCoords).toEqual({rowIndex: -10, visibleColumnIndex: 14});
+            // when grid has no data
+            grid.filter('col0', 2, IgxNumberFilteringOperand.instance().condition('greaterThan'));
+            fix.detectChanges();
+
+            nextCellCoords = grid.getNextCell(0, 0, (col) => col.editable);
+            expect(nextCellCoords).toEqual({rowIndex: 0, visibleColumnIndex: 0});
+        });
+
+        it(`GetPreviousCell: should return correctly next cell coordinates`, async() => {
+            const fix = TestBed.createComponent(IgxGridDefaultRenderingComponent);
+            fix.componentInstance.initColumnsRows(15, 5);
+            fix.detectChanges();
+
+            const grid = fix.componentInstance.grid;
+            grid.height = '500px';
+            await wait(30);
+            fix.detectChanges();
+
+            grid.getColumnByName('col2').editable = true;
+            fix.detectChanges();
+            grid.getColumnByName('col4').editable = true;
+            fix.detectChanges();
+            // when the previous cell is on the same row
+            let prevCellCoords = grid.getPreviousCell(0, 4, (col) => col.editable);
+            expect(prevCellCoords).toEqual({rowIndex: 0, visibleColumnIndex: 2});
+            // when the previous cell is on the previous row
+            prevCellCoords = grid.getPreviousCell(1, 2, (col) => col.editable);
+            expect(prevCellCoords).toEqual({rowIndex: 0, visibleColumnIndex: 4});
+            // when the current row and column index are not valid
+            prevCellCoords = grid.getPreviousCell(-110, 2, (col) => col.editable);
+            expect(prevCellCoords).toEqual({rowIndex: -110, visibleColumnIndex: 2});
+            // when there is no previous cell
+            prevCellCoords = grid.getPreviousCell(0, 2, (col) => col.editable);
+            expect(prevCellCoords).toEqual({rowIndex: 0, visibleColumnIndex: 2});
+             // when the filter function has no matching colums
+             prevCellCoords = grid.getPreviousCell(0, 3, (col) => col.movable);
+             expect(prevCellCoords).toEqual({rowIndex: 0, visibleColumnIndex: 3});
+            // when grid has no data
+            grid.filter('col0', 2, IgxNumberFilteringOperand.instance().condition('greaterThan'));
+            fix.detectChanges();
+
+            prevCellCoords = grid.getPreviousCell(99, 0, (col) => col.editable);
+            expect(prevCellCoords).toEqual({rowIndex: 99, visibleColumnIndex: 0});
+        });
 
         it(`Should not commit added row to grid's data in grid with transactions`, fakeAsync(() => {
             const fixture = TestBed.createComponent(IgxGridRowEditingTransactionComponent);
@@ -2433,7 +2597,7 @@ describe('IgxGrid Component Tests', () => {
                 expect(cell.value).toBe(110); // SORT does not submit
 
                 expect(grid.endEdit).toHaveBeenCalled();
-                expect(grid.endEdit).toHaveBeenCalledWith(true);
+                expect(grid.endEdit).toHaveBeenCalledWith(false);
                 expect(cell.inEditMode).toBeFalsy();
             }));
 
@@ -2658,9 +2822,8 @@ describe('IgxGrid Component Tests', () => {
                 expect(grid.rowEditingOverlay.collapsed).toBeTruthy();
             }));
 
-            it(`Should show the updated value when showing the column again`, fakeAsync(() => {
+            it(`Should show the previous value when showing the column again`, fakeAsync(() => {
                 const targetCbText = 'Product Name';
-                const newValue = 'Tea';
                 const fix = TestBed.createComponent(IgxGridRowEditingComponent);
                 fix.detectChanges();
 
@@ -2668,7 +2831,7 @@ describe('IgxGrid Component Tests', () => {
                 const targetCell = grid.getCellByColumn(0, 'ProductName');
                 targetCell.inEditMode = true;
                 tick();
-                targetCell.update(newValue);
+                targetCell.update('Tea');
 
                 // hide column
                 grid.toolbar.columnHidingButton.nativeElement.click();
@@ -2681,7 +2844,7 @@ describe('IgxGrid Component Tests', () => {
                 targetCheckbox.nativeElement.click();
                 tick();
 
-                expect(targetCell.value).toEqual(newValue);
+                expect(targetCell.value).toEqual('Chai');
             }));
 
             xit(`Should be possible to update a cell that is hidden programmatically`, () => { // This is NOT possible
@@ -2919,6 +3082,27 @@ describe('IgxGrid Component Tests', () => {
         });
 
         describe('Row Editing - Transaction', () => {
+            it('Should properly exit pending state when committing row edit w/o changes', fakeAsync(() => {
+                const fixture = TestBed.createComponent(IgxBasicGridRowEditingComponent);
+                fixture.detectChanges();
+
+                const grid = fixture.componentInstance.grid;
+                const initialDataLength = grid.data.length;
+                const productNameCell = fixture.debugElement.queryAll(By.css('.igx-grid__td'))[2];
+                const enterEvent = { key: 'enter', stopPropagation: () => {}, preventDefault: () => {} };
+                productNameCell.triggerEventHandler('keydown', enterEvent);
+                tick();
+                fixture.detectChanges();
+                expect(grid.getCellByKey(1, 'ProductName').inEditMode).toBeTruthy();
+                productNameCell.triggerEventHandler('keydown', enterEvent);
+                tick();
+                fixture.detectChanges();
+                expect(grid.getCellByKey(1, 'ProductName').inEditMode).toBeFalsy();
+                grid.deleteRow(2);
+                tick();
+                fixture.detectChanges();
+                expect(grid.data.length).toEqual(initialDataLength - 1);
+            }));
             it('Transaction Update, Delete, Add, Undo, Redo, Commit check transaction and grid state', fakeAsync(() => {
                 const fixture = TestBed.createComponent(IgxGridRowEditingTransactionComponent);
                 fixture.detectChanges();
@@ -3701,6 +3885,21 @@ export class IgxGridMarkupDeclarationComponent extends IgxGridTestComponent {
     ];
     @ViewChild(IgxGridComponent, { read: IgxGridComponent })
     public instance: IgxGridComponent;
+}
+
+@Component({
+    template: `<div>
+        <igx-grid [data]="data" (onColumnInit)="columnCreated($event)">
+            <igx-column field="ID"></igx-column>
+            <igx-column field="Name"></igx-column>
+        </igx-grid>
+        </div>
+    `
+})
+export class IgxGridEmptyMessage100PercentComponent extends IgxGridTestComponent {
+    public data = [];
+    @ViewChild(IgxGridComponent, { read: IgxGridComponent })
+    public grid: IgxGridComponent;
 }
 
 @Injectable()