+++ CHANGELOG.md
@@ -2,9 +2,26 @@
 
 All notable changes for each version of this project will be documented in this file.
 
-## 7.2.6
+## 7.2.x
 ### New feature
-- **igxSlider** - exposing new `labels` property accepting a collection of literal values that become equally spread over the slider, by placing each element as a thumb label. 
+- **igxSlider** - exposing new `labels` property accepting a collection of literal values that become equally spread over the slider, by placing each element as a thumb label.
+ 
+## 7.2.9
+- `Pager`
+    - **Behavioral Change** - The pager is now hidden when there are no records in the grid.
+
+## 7.2.8
+- `IgxGrid` Custom keyboard navigation
+    - `onFocusChange` event is deprecated.
+    - `onGridKeydown` is exposed. The event will emit
+    `IGridKeydownEventArgs { targetType: GridKeydownTargetType; target: Object; event: Event; cancel: boolean; }`
+    - `navigateTo(rowIndex: number, visibleColumnIndex: number, callback({targetType, target: Object }))` - this method allows you to navigate to a position in the grid based on provided `rowindex` and `visibleColumnIndex`;
+    - `getNextCell(currentRowIndex, currentvisibleColumnIndex, callback(IgxColumnComponent))` - returns `{ rowIndex, visibleColumnIndex }` which defines the next cell, that match specific criteria according to the current position
+    - `getPreviousCell(currentRowIndex, currentvisibleColumnIndex, callback(IgxColumnComponent))` - returns `{ rowIndex, visibleColumnIndex }` which defines the previous cell, that match specific criteria according to the current position
+
+## 7.2.6
+- `igxGrid`
+    - **Feature** The `groupsRecords` property now returns the full grouping tree as in 7.1 and also includes the grouping information for all pages.
 
 ## 7.2.5
 - `igxDrop`
@@ -40,9 +57,9 @@ All notable changes for each version of this project will be documented in this 
 ### New feature
 - [Multi-cell selection](https://github.com/IgniteUI/igniteui-angular/wiki/Grid-Multi-cell-selection-Specification) - Enables range selection of cells in the grid.
 
-### Grids Performance improvements 
+### Grids Performance improvements
 - Grid rendering speed
-- Grid grouping rendering speed 
+- Grid grouping rendering speed
 - Grid vertical scrolling using the scroll arrows
 - Grid horizontal scrolling using the scroll arrows
 - Grid cell focusing time
@@ -81,7 +98,7 @@ All notable changes for each version of this project will be documented in this 
 - Update child summaries correctly when CRUD operations are performed #4408
 - Add igxQuickFilterTemplate directive #4377
 - Resizing: move resize handle logic in a directive #4378
-- No event emitted when column is unpinned #3799 
+- No event emitted when column is unpinned #3799
 - When update a cell in the grouped column the child summaries are not updated #4324
 - Column Group border is misaligned with its children's in some cases #4387
 - Expanding last row of HierarchicalGrid via keyboard(Alt + downArrow) leads to cell losing its focus. #4080
@@ -92,7 +109,7 @@ All notable changes for each version of this project will be documented in this 
 ## 7.2.2
 ### Features
 - **Components' Display Type** - All components now have their CSS display property explicitly set on the host element to ensure width, padding, and margins are applied when set directly on the host selectors.
-- **Themes** 
+- **Themes**
     - Add support for gradients and images as values for component themes via the component theme functions.
     - `Palettes` - added surface color to the palette. The surface color is used by cards, pickers, dialog windows, etc. as the default background.
 
@@ -317,13 +334,13 @@ All notable changes for each version of this project will be documented in this 
 - IgxCombo - Keyboard navigation ArrowDown stutters on chunk load #3999
 - Row editing overlay banner not shown when enter row editing #4117
 - IgxToggle open method always tries to get id even when it has one #3971
-- Last (right-aligned) column is cut off when no widths are set for the columns #3396 
+- Last (right-aligned) column is cut off when no widths are set for the columns #3396
 - The selection in the last grid column does not span in the whole cell. #1115
 - Last column header is a bit wider than the cells #1230
 
 ## 7.1.11
 ### Improvements
-- Row and Cell editing Docs improvements #4055 
+- Row and Cell editing Docs improvements #4055
 
 ## 7.1.10
 ### Features